+++ src/Ext/SWType/Body.h
@@ -24,6 +24,14 @@ public:
 
 		//Ares 0.A
 		Valueable<int> Money_Amount;
+		ValueableIdx<VoxClass> EVA_Impatient;
+		ValueableIdx<VoxClass> EVA_InsufficientFunds;
+		ValueableIdx<VoxClass> EVA_SelectTarget;
+		Valueable<bool> SW_UseAITargeting;
+		Valueable<bool> SW_AutoFire;
+		Valueable<bool> SW_ManualFire;
+		Valueable<bool> SW_ShowCameo;
+		Valueable<bool> SW_Unstoppable;
 		ValueableVector<TechnoTypeClass*> SW_Inhibitors;
 		Valueable<bool> SW_AnyInhibitor;
 		ValueableVector<TechnoTypeClass*> SW_Designators;
@@ -39,6 +47,11 @@ public:
 		Valueable<bool> SW_InitialReady;
 		ValueableIdx<SuperWeaponTypeClass> SW_PostDependent;
 		Valueable<int> SW_MaxCount;
+
+		Valueable<CSFText> Message_CannotFire;
+		Valueable<CSFText> Message_InsufficientFunds;
+		Valueable<int> Message_ColorScheme;
+		Valueable<bool> Message_FirerColor;
 
 		Valueable<CSFText> UIDescription;
 		Valueable<int> CameoPriority;
@@ -64,6 +77,13 @@ public:
 		Valueable<bool> ShowDesignatorRange;
 
 		Valueable<int> TabIndex;
+        
+		Valueable<bool> ExclusiveSidebar_Allow;
+		DWORD ExclusiveSidebar_PriorityHouses;
+		DWORD ExclusiveSidebar_RequiredHouses;
+
+		CustomPalette SidebarPal;
+		PhobosPCXFile SidebarPCX;
 
 		std::vector<ValueableVector<int>> LimboDelivery_RandomWeightsData;
 		std::vector<ValueableVector<int>> SW_Next_RandomWeightsData;
@@ -82,6 +102,14 @@ public:
 
 		ExtData(SuperWeaponTypeClass* OwnerObject) : Extension<SuperWeaponTypeClass>(OwnerObject)
 			, Money_Amount { 0 }
+			, EVA_Impatient { -1 }
+			, EVA_InsufficientFunds { -1 }
+			, EVA_SelectTarget { -1 }
+			, SW_UseAITargeting { false }
+			, SW_AutoFire { false }
+			, SW_ManualFire { true }
+			, SW_ShowCameo { true }
+			, SW_Unstoppable { false }
 			, SW_Inhibitors {}
 			, SW_AnyInhibitor { false }
 			, SW_Designators { }
@@ -96,6 +124,10 @@ public:
 			, SW_PostDependent {}
 			, SW_MaxCount { -1 }
 			, SW_Shots { -1 }
+			, Message_CannotFire {}
+			, Message_InsufficientFunds {}
+			, Message_ColorScheme { -1 }
+			, Message_FirerColor { false }
 			, UIDescription {}
 			, CameoPriority { 0 }
 			, LimboDelivery_Types {}
@@ -120,6 +152,11 @@ public:
 			, Convert_Pairs {}
 			, ShowDesignatorRange { true }
 			, TabIndex { 1 }
+			, ExclusiveSidebar_Allow { true }
+			, ExclusiveSidebar_PriorityHouses { 0u }
+			, ExclusiveSidebar_RequiredHouses { 0xFFFFFFFFu }
+			, SidebarPal {}
+			, SidebarPCX {}
 			, UseWeeds { false }
 			, UseWeeds_Amount { RulesClass::Instance->WeedCapacity }
 			, UseWeeds_StorageTimer { false }
@@ -141,6 +178,7 @@ public:
 		bool IsLaunchSite(BuildingClass* pBuilding) const;
 		std::pair<double, double> GetLaunchSiteRange(BuildingClass* pBuilding = nullptr) const;
 		bool IsAvailable(HouseClass* pHouse) const;
+		void PrintMessage(const CSFText& message, HouseClass* pFirer) const;
 
 		void ApplyLimboDelivery(HouseClass* pHouse);
 		void ApplyLimboKill(HouseClass* pHouse);