+++ src/Ext/Side/Body.cpp
@@ -42,6 +42,11 @@ void SideExt::ExtData::LoadFromINIFile(CCINIClass* pINI)
 	this->ToolTip_Background_Opacity.Read(exINI, pSection, "ToolTip.Background.Opacity");
 	this->ToolTip_Background_BlurSize.Read(exINI, pSection, "ToolTip.Background.BlurSize");
 	this->BriefingTheme = pINI->ReadTheme(pSection, "BriefingTheme", this->BriefingTheme);
+
+	this->ExclusiveSWSidebar_TopPCX.Read(pINI, pSection, "ExclusiveSWSidebar.TopPCX");
+	this->ExclusiveSWSidebar_CenterPCX.Read(pINI, pSection, "ExclusiveSWSidebar.CenterPCX");
+	this->ExclusiveSWSidebar_BottomPCX.Read(pINI, pSection, "ExclusiveSWSidebar.BottomPCX");
+	this->ExclusiveSWSidebar_ToggleShape.Read(exINI, pSection, "ExclusiveSWSidebar.ToggleShape");
 }
 
 // =============================
@@ -71,6 +76,10 @@ void SideExt::ExtData::Serialize(T& Stm)
 		.Process(this->IngameScore_WinTheme)
 		.Process(this->IngameScore_LoseTheme)
 		.Process(this->BriefingTheme)
+		.Process(this->ExclusiveSWSidebar_TopPCX)
+		.Process(this->ExclusiveSWSidebar_CenterPCX)
+		.Process(this->ExclusiveSWSidebar_BottomPCX)
+		.Process(this->ExclusiveSWSidebar_ToggleShape)
 		;
 }